Break table of budget records out into its own component for tidiness
